[BUGFIX] use cache hash when building product link in solr view helper
